+++ test/utest-online/testlinkapi_online_test.py
@@ -478,6 +478,10 @@ class TestLinkAPIOnlineTestCase(unittest.TestCase):
             self.client.setTestCaseTestSuite(
                             'TC-40000712', '40000713 suite')
          
+    def test_getTestSuiteAttachments_unknownID(self):
+        with self.assertRaisesRegex(TLResponseError, '8000.*40000712'):
+            self.client.getTestSuiteAttachments(40000712)
+         
         
 if __name__ == "__main__":
     #import sys;sys.argv = ['', 'Test.testName']